metodo para somar limite dos cartoes por portador criado

+++ src/main/java/com/client/portador/apicreditanalysis/dto/CreditAnalysisDto.java
@@ -1,16 +1,17 @@
 package com.client.portador.apicreditanalysis.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 import lombok.Builder;
 
 public record CreditAnalysisDto(
         UUID id,
         UUID clientId,
-        Double approvedLimit,
+        BigDecimal approvedLimit,
         Boolean approved
 ) {
     @Builder(toBuilder = true)
-    public CreditAnalysisDto(UUID id, UUID clientId, Double approvedLimit, Boolean approved) {
+    public CreditAnalysisDto(UUID id, UUID clientId, BigDecimal approvedLimit, Boolean approved) {
         this.id = id;
         this.clientId = clientId;
         this.approvedLimit = approvedLimit;
