+++ src/main/java/com/client/portador/controller/response/PortadorResponse.java
@@ -1,12 +1,13 @@
 package com.client.portador.controller.response;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
 public record PortadorResponse(
         UUID cardHolderId,
         String status,
-        Double limit,
+        BigDecimal limit,
         LocalDateTime createdAt
 
 ) {
