+++ src/main/java/com/client/portador/model/Portador.java
@@ -1,6 +1,7 @@
 package com.client.portador.model;
 
 import com.client.portador.utils.Status;
+import java.math.BigDecimal;
 import java.util.UUID;
 import lombok.Builder;
 
@@ -8,11 +9,11 @@ public record Portador(
         Status status,
         UUID clientId,
         UUID creditAnalysisId,
-        Double limit,
+        BigDecimal limit,
         BankAccount bankAccount
 ) {
     @Builder(toBuilder = true)
-    public Portador(Status status, UUID clientId, UUID creditAnalysisId, Double limit, BankAccount bankAccount) {
+    public Portador(Status status, UUID clientId, UUID creditAnalysisId, BigDecimal limit, BankAccount bankAccount) {
         this.status = status;
         this.limit = limit;
         this.creditAnalysisId = creditAnalysisId;
@@ -20,7 +21,7 @@ public record Portador(
         this.bankAccount = bankAccount;
     }
 
-    public Portador updateLimiteFrom(Double limit) {
+    public Portador updateLimiteFrom(BigDecimal limit) {
         return this.toBuilder()
                 .status(Status.ATIVO)
                 .limit(limit).build();
